Feat: Логика для GET запросов

+++ src/services/equipment.service.ts
@@ -1,27 +1,29 @@
 import { Equipment } from "../domains/entities/equipment.entity.js";
-import { CreateEquipmentRequest } from "../dto/contracts/equipment/create-equipment.request.js";
-import { GetEquipmentWeatherResponse } from "../dto/contracts/equipment/get-equipment-weather.response.js";
-import { getEquipmentsRequest } from "../dto/contracts/equipment/get-equipments.request.js";
-import { UpdateEquipmentRequest } from "../dto/contracts/equipment/update-equipment.request.js";
+import { CreateEquipmentRequest } from "../dto/equipment/create-equipment.request.js";
+import { GetEquipmentWeatherResponse } from "../dto/equipment/get-equipment-weather.response.js";
+import { getEquipmentsRequest } from "../dto/equipment/get-equipments.request.js";
+import { UpdateEquipmentRequest } from "../dto/equipment/update-equipment.request.js";
+import { ConflictError } from "../errors/conflicts.error.js";
+import { NotFoundError } from "../errors/not-found.error.js";
 import { IEquipmentRepository } from "../repositories/abstractions/equipment-repository.interface.js";
+import { IMaintenanceRequestRepository } from "../repositories/abstractions/maintenance-request.repository.interface.js";
 import { EquipmentRepositoryMemory } from "../repositories/implementations/in-memory-equipment-repository.js";
 import { weatherSuitableSchema } from "../validators/schemas/equipment/weather-suitable.schema.js";
 import { getWeatherAsync } from "./weather.service.js";
+import { MaintenanceRequestRepository } from './../repositories/implementations/in-memory-maintenance-request.repository.js';
+import { MaintenanceRequestStatus } from "../domains/enums/maintenance-request-status.enum.js";
+import { GetEquipmentsResult } from "../dto/equipment/get-equipments.result.js";
+import { getRequests } from "./maintenance-request.service.js";
+import { GetEquipmentsRequests } from "../dto/equipment/get-equipment-requests.request.js";
+import { MaintenanceRequest } from "../domains/entities/maintenance-request.entity.js";
 
 const repository: IEquipmentRepository = new EquipmentRepositoryMemory();
+const requestsRepository: IMaintenanceRequestRepository = new MaintenanceRequestRepository();
 
-//todo :rewrite errors
 export const addEquipment = async(equipmentInfo: CreateEquipmentRequest): Promise<string> => {
-
-  const installedDate = new Date(equipmentInfo.installedAt);
-  const currentDate = new Date();
-  if (installedDate > currentDate) {
-    throw new Error("Дата в будущем");
-  }
-
   const existing = await repository.getBySerialNumber(equipmentInfo.serialNumber);
   if (existing) {
-    throw new Error("Сериал номер уже есть");
+    throw new ConflictError("Оборудование с указанным серийным номером уже существует!", [{field: "serialNumber", message: "Неуникальный серийный номер"}]);
   }
   
   const equipment = Equipment.create(equipmentInfo);
@@ -29,14 +31,25 @@ export const addEquipment = async(equipmentInfo: CreateEquipmentRequest): Promis
   return equipment.id;
 };
 
-export const getEquipments = async(request: getEquipmentsRequest): Promise<Equipment[]> => {
-  return await repository.get(request);
+export const getEquipments = async(request: getEquipmentsRequest): Promise<GetEquipmentsResult> => {
+  const serviceResult = new GetEquipmentsResult();
+
+  const result = await repository.get(request);
+  serviceResult.equipments = result.equipments;
+  serviceResult.total = result.total;
+
+  return serviceResult;
 };
 
 export const deleteEquipment = async(id: string): Promise<void> => {
   const existing = await repository.getById(id);
   if (!existing) {
-    throw new Error("Не найден");
+    throw new NotFoundError("Оборудование не найдено!", [{field: "id", message: `Оборудования с id = ${id} не существует`}]);
+  }
+
+  const hasUnfinishedRequests = await requestsRepository.existWithStatuses(id, [MaintenanceRequestStatus.new, MaintenanceRequestStatus.in_progress]);
+  if (hasUnfinishedRequests) {
+    throw new ConflictError("Для данного оборудования есть незавершенные заявки", [{field: "id", message: `Оборудования с id = ${id} имеет открытыие заявки`}]);
   }
 
   await repository.delete(id);
@@ -45,7 +58,7 @@ export const deleteEquipment = async(id: string): Promise<void> => {
 export const getEquipment = async(id: string): Promise<Equipment | undefined> => {
   const existing = await repository.getById(id);
   if (!existing) {
-    throw new Error("Не найден");
+    throw new NotFoundError("Оборудование не найдено!", [{field: "id", message: `Оборудования с id = ${id} не существует`}]);
   }
 
   return await repository.getById(id);
@@ -54,7 +67,7 @@ export const getEquipment = async(id: string): Promise<Equipment | undefined> =>
 export const updateEquipment = async(id: string, equipmentInfo: UpdateEquipmentRequest): Promise<void> => {
   const equipment = await repository.getById(id);
   if (!equipment) {
-    throw new Error("Не найден");
+     throw new NotFoundError("Оборудование не найдено!", [{field: "id", message: `Оборудования с id = ${id} не существует`}]);
   }
 
   if (equipment) {
@@ -68,7 +81,7 @@ export const getEquipmentWeather = async(id: string): Promise<GetEquipmentWeathe
   const response = new GetEquipmentWeatherResponse();
   const equipment = await repository.getById(id);
   if (!equipment) {
-    throw new Error("Не найден");
+     throw new NotFoundError("Оборудование не найдено!", [{field: "id", message: `Оборудования с id = ${id} не существует`}]);
   }
 
   response.equipmentId = id;
@@ -84,4 +97,14 @@ export const getEquipmentWeather = async(id: string): Promise<GetEquipmentWeathe
   response.isWeatherWindowSuitable = response.weather.every(x => x.suitable);
 
   return response;
+}
+
+export const getEquipmentsMaintenanceRequests = async(id: string, request: GetEquipmentsRequests): Promise<MaintenanceRequest[]> => {
+  const equipment = await repository.getById(id);
+  if (!equipment) {
+     throw new NotFoundError("Оборудование не найдено!", [{field: "id", message: `Оборудования с id = ${id} не существует`}]);
+  }
+
+  let maintenanceRequest = {...request, equipmentsId: [id]};
+  return await getRequests(maintenanceRequest);
 }
